D8-165 Added floating form data to appointments page
