+++ package.json
@@ -12,7 +12,7 @@
     "@bradgarropy/next-google-analytics": "^1.0.2",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.14.8",
+    "@mui/material": "^5.14.9",
     "@sendgrid/mail": "^7.7.0",
     "dotenv": "^16.0.3",
     "eslint": "8.36.0",
